cleanup, change default max shards to 100 from 1000 & more

+++ coerce/tests/test_actor_messaging.rs
@@ -126,7 +126,7 @@ pub async fn test_actor_notify() {
         .await
         .unwrap();
 
-    for i in 1..=25 as i32 {
+    for i in 1..=25_i32 {
         let _ = actor_ref.notify_exec(move |mut actor| actor.counter = i);
     }
